+++ src/app/organisms/candidacy-overview/candidacy-overview.component.ts
@@ -4,40 +4,33 @@
  * For license information see LICENSE file.
  */
 
-import {
-  DialogService,
-  PaginatorComponent,
-  SelectionChange,
-  SortDirective,
-  TableDataSource,
-} from '@abraxas/base-components';
+import { DialogService, SelectionChange, SortDirective, TableDataSource, } from '@abraxas/base-components';
 import {
   AfterViewInit,
   Component,
   EventEmitter,
+  inject,
   Input,
   OnDestroy,
   OnInit,
   Output,
   ViewChild,
-  inject,
 } from '@angular/core';
 import { TranslateService } from '@ngx-translate/core';
 import { forkJoin, Subscription } from 'rxjs';
 import { finalize } from 'rxjs/operators';
 import { GuardService } from 'src/app/shared/guard.service';
-import { CandidateModel } from 'src/app/shared/models/candidate.model';
-import { ListCandidateModel, newListCandidateModel } from 'src/app/shared/models/ListCandidate.model';
+import { CandidateModel, newCandidateModel } from 'src/app/shared/models/candidate.model';
 import { CandidateService } from 'src/app/shared/services/candidate.service';
 import { RxJsUtilsService } from 'src/app/shared/services/rx-js-utils.service';
 import { SnackbarService } from 'src/app/shared/services/snackbar.service';
-import { v4 as uuid } from 'uuid';
 import { ElectionModel, ElectionType } from '../../shared/models/election.model';
 import { ListModel } from '../../shared/models/list.model';
 import { SettingsModel } from '../../shared/models/settings.model';
 import { SettingsService } from '../../shared/services/settings.service';
 import { CandidacyDetailsComponent } from '../candidacy-details/candidacy-details.component';
 import { CandidacyModifyComponent } from '../candidacy-modify/candidacy-modify.component';
+import { ListCandidateModel } from '../../shared/models/list-candidate.model';
 
 @Component({
   selector: 'app-candidacy-overview',
@@ -53,14 +46,15 @@ export class CandidacyOverviewComponent implements OnInit, OnDestroy, AfterViewI
   private settingsService = inject(SettingsService);
   private dialogService = inject(DialogService);
   private translateService = inject(TranslateService);
+  private readonly candidacyDialogWidth = '45rem';
 
   @ViewChild('candidateSort') public candidateSort: SortDirective;
 
-  public candidates: ListCandidateModel[] = [];
-  public initialCandidacies: ListCandidateModel[] = [];
+  public expandedCandidates: ListCandidateModel[] = [];
+  public candidates: CandidateModel[] = [];
+  public initialCandidates: CandidateModel[] = [];
   public selected?: ListCandidateModel;
   public hasChanges: boolean = false;
-  public candidate?: CandidateModel;
   public canEdit: boolean = false;
   public loading: boolean = false;
   public saving: boolean = false;
@@ -71,7 +65,7 @@ export class CandidacyOverviewComponent implements OnInit, OnDestroy, AfterViewI
   @Input() public list: ListModel;
   @Output() public hasChangesChange: EventEmitter<boolean> = new EventEmitter<boolean>();
 
-  public dataSource = new TableDataSource<ListCandidateModel>(this.candidates);
+  public dataSource = new TableDataSource<ListCandidateModel>(this.expandedCandidates);
   public columns = TableColumn;
   public columnsToDisplay: string[] = [
     TableColumn.icons,
@@ -98,6 +92,10 @@ export class CandidacyOverviewComponent implements OnInit, OnDestroy, AfterViewI
       .subscribe((res) => {
         this.canEdit = !this.election.isArchived ? res : false;
       });
+
+    if (this.election.electionType === ElectionType.Majorz) {
+      this.columnsToDisplay = this.columnsToDisplay.filter((c) => c !== TableColumn.number);
+    }
 
     this.loading = true;
     const candidates$ = this.candidateService.getAll(this.election.id, this.list.id);
@@ -132,12 +130,12 @@ export class CandidacyOverviewComponent implements OnInit, OnDestroy, AfterViewI
   }
 
   public get hasNrOfMandatesReached(): boolean {
-    return !this.candidates || this.nrOfMandates <= this.candidates.length;
+    return this.nrOfMandates <= this.expandedCandidates.length;
   }
 
   public saveChanges(): void {
     this.saving = true;
-    const submitCandidates = this.candidates.map((c) => {
+    const candidatesToSave = this.candidates.map((c) => {
       return {
         ...c,
         ballotFamilyName: c.ballotFamilyName || c.familyName,
@@ -147,21 +145,7 @@ export class CandidacyOverviewComponent implements OnInit, OnDestroy, AfterViewI
       };
     });
 
-    const clonedCandidates: ListCandidateModel[] = [];
-    // Remove clones from the list
-    submitCandidates
-      .filter((c) => c.cloned)
-      .forEach((c) => {
-        const index = submitCandidates.findIndex((c2) => c.uuid === c2.uuid);
-        const submitCandidateToRemove = submitCandidates[index];
-        submitCandidates.splice(index, 1);
-
-        if (submitCandidateToRemove.isOriginal) {
-          clonedCandidates.push(submitCandidateToRemove);
-        }
-      });
-    let candidatesToSave = [...submitCandidates, ...clonedCandidates];
-    candidatesToSave.sort((a, b) => this.sortCandidateAscendingWithIndexAndOrderIndex(a, b));
+    candidatesToSave.sort((a, b) => a.index - b.index);
     this.candidateService
       .batchUpdate(this.election.id, this.list.id, candidatesToSave)
       .pipe(
@@ -171,8 +155,8 @@ export class CandidacyOverviewComponent implements OnInit, OnDestroy, AfterViewI
       .subscribe((cs) => {
         this.processCandidates(cs);
         this.refreshExpandedCandidates();
-        this.updateHasChanges();
         this.hasChanges = false;
+        this.hasChangesChange.emit(this.hasChanges);
         this.snackbarService.success(this.translateService.instant('CANDIDACY.MSG_LIST_UPDATE_SUCCESS'));
       });
   }
@@ -185,141 +169,109 @@ export class CandidacyOverviewComponent implements OnInit, OnDestroy, AfterViewI
     }
   }
 
-  public createCandidate(candidacy: ListCandidateModel, nextRequested: boolean): void {
-    candidacy.orderIndex = this.candidates.length + 1;
-    this.candidates.push({
-      ...candidacy,
-    });
+  public createCandidate(candidacy: CandidateModel): void {
+    candidacy.orderIndex = this.expandedCandidates.length + 1;
+    candidacy.index = candidacy.orderIndex;
+    this.candidates.push(candidacy);
     this.refreshExpandedCandidates();
-    this.dataSource.data = this.candidates;
-    if (!nextRequested) {
-      this.candidate = null;
-    }
     this.saveChanges();
   }
 
-  public modifyCandidate(candidacy: ListCandidateModel, nextRequested: boolean): void {
-    this.candidates = this.candidates.map((c) => {
-      if (c.uuid === candidacy.uuid) {
-        return {
-          ...c,
-          ...candidacy,
-        };
-      }
-      if (c.uuid === candidacy.cloneUuid) {
-        return {
-          ...c,
-          ...candidacy,
-          uuid: c.uuid,
-          cloneUuid: c.cloneUuid,
-        };
-      }
-      return c;
-    });
-    this.refreshExpandedCandidates();
-    this.dataSource.data = this.candidates;
-    if (!nextRequested) {
-      this.candidate = null;
-    }
-    this.updateHasChanges();
-  }
-
-  public isTagged(candidate: CandidateModel, field: string): boolean {
-    return candidate.markings.findIndex((f) => f.field === field) !== -1;
+  public isTagged(candidate: ListCandidateModel, field: string): boolean {
+    return candidate.candidate.markings.findIndex((f) => f.field === field) !== -1;
   }
 
   public addCandidacy() {
-    const newcandidate = newListCandidateModel();
+    const newCandidate = newCandidateModel(this.expandedCandidates.length + 1);
     let dialogRef = this.dialogService.open(CandidacyModifyComponent, {
       data: {
         election: this.election,
         settings: this.settings,
-        candidacy: newcandidate,
+        candidacy: newCandidate,
         maxCandidateCount: this.nrOfMandates,
-        candidateCount: this.candidates.length,
+        candidateCount: this.expandedCandidates.length,
       },
-    });
+    },
+      this.candidacyDialogWidth);
 
     dialogRef.componentInstance.formSubmit.subscribe((data) => {
-      if (!this.hasNrOfMandatesReached) {
-        this.createCandidate(data.candidacy, data.nextRequested);
-      } else {
-        this.snackbarService.warning(
-          this.translateService.instant('CANDIDACY.REACHED_MAXIMUM_CANDIDATES_WARNING')
-        );
-      }
+      this.handleCandidateCreated(data.candidacy);
     });
   }
 
   public modifyCandidacy(candidate: ListCandidateModel): void {
-    let dialogRef = this.dialogService.open(CandidacyModifyComponent, {
-      data: {
-        election: this.election,
-        settings: this.settings,
-        candidacy: candidate,
-        maxCandidateCount: this.nrOfMandates,
-        candidateCount: this.candidates.length,
+    const dialogRef = this.dialogService.open(
+      CandidacyModifyComponent,
+      {
+        data: {
+          election: this.election,
+          settings: this.settings,
+          candidacy: { ...candidate.candidate, markings: [...candidate.candidate.markings] },
+          maxCandidateCount: this.nrOfMandates,
+          candidateCount: this.expandedCandidates.length,
+        },
       },
-    });
-    dialogRef.componentInstance.formSubmit.subscribe((data) => {
-      this.modifyCandidate(data.candidacy, data.nextRequested);
+      this.candidacyDialogWidth
+    );
+    dialogRef.componentInstance.formSubmit.subscribe(({ candidacy }) => {
+      if (candidacy.index === candidate.candidate.index) {
+        this.candidates[candidacy.index - 1] = candidacy;
+        this.refreshExpandedCandidates();
+        this.updateHasChanges();
+      } else {
+        // New candidate created via the "next" button of the dialog
+        this.handleCandidateCreated(candidacy);
+      }
     });
   }
 
+  private handleCandidateCreated(candidate: CandidateModel): void {
+    if (this.hasNrOfMandatesReached) {
+      this.snackbarService.warning(this.translateService.instant('CANDIDACY.REACHED_MAXIMUM_CANDIDATES_WARNING'));
+      return;
+    }
+
+    this.createCandidate(candidate);
+  }
+
   public showDetailsModal(candidate: ListCandidateModel): void {
-    this.candidate = candidate;
     this.dialogService.open(CandidacyDetailsComponent, {
       data: {
         election: this.election,
         settings: this.settings,
-        candidacy: this.candidate,
+        candidacy: candidate.candidate,
         listId: this.list.id,
       },
     });
   }
 
   public remove(): void {
-    const index = this.candidates.indexOf(this.selected);
-    this.candidates
-      .filter((c) => c.cloneUuid === this.candidates[index].uuid)
-      .forEach((c) => (c.cloned = false));
-    this.candidates.splice(index, 1);
-    this.candidates.forEach((c) => {
-      if (!c.cloned) {
-        delete c.cloneOrderIndex;
-      }
-    });
+    if (this.selected.isCloned) {
+      this.selected.candidate.cloned = false;
+      delete this.selected.candidate.cloneOrderIndex;
+    } else if (this.selected.candidate.cloned) {
+      this.selected.candidate.cloned = false;
+      this.selected.candidate.index = this.selected.candidate.cloneOrderIndex;
+      this.selected.candidate.orderIndex = this.selected.candidate.cloneOrderIndex;
+      delete this.selected.candidate.cloneOrderIndex;
+    } else {
+      this.candidates = this.candidates.filter((c) => c !== this.selected.candidate);
+    }
+
     this.refreshExpandedCandidates();
     this.updateCandidatePositions();
-    this.dataSource.data = this.candidates;
     this.updateHasChanges();
   }
 
   public preCumulate(): void {
-    if (!this.selected || this.selected.cloned) {
+    if (!this.selected || this.selected.candidate.cloned) {
       return;
     }
 
-    const candidateIndex = this.candidates.findIndex((c2) => c2.uuid === this.selected.uuid);
-    const candidate = this.candidates[candidateIndex];
-    candidate.cloned = true;
-    const newCandidate = { ...candidate };
-    newCandidate.uuid = uuid();
-    candidate.cloneUuid = newCandidate.uuid;
-    newCandidate.cloneUuid = candidate.uuid;
-    newCandidate.isOriginal = false;
-    newCandidate.cloneOrderIndex = candidate.orderIndex + 1;
-    candidate.cloneOrderIndex = candidate.orderIndex + 1;
-    this.candidates.splice(candidateIndex + 1, 0, newCandidate);
-    this.updateCandidatePositions();
-    this.dataSource.data = this.candidates;
-    this.updateHasChanges();
-  }
-
-  public reorderCandidates(updatedList: ListCandidateModel[]): void {
-    this.candidates = updatedList;
-    this.updateCandidatePositions();
+    this.selected.candidate.cloned = true;
+    this.selected.candidate.cloneOrderIndex = this.expandedCandidates.length + 1;
     this.refreshExpandedCandidates();
+    this.updateHasChanges();
   }
 
   public async moveCandidate(previousIndex: number, newIndex: number): Promise<void> {
@@ -327,117 +279,55 @@ export class CandidacyOverviewComponent implements OnInit, OnDestroy, AfterViewI
       return;
     }
 
-    const movedCandidate = this.candidates.splice(previousIndex, 1)[0];
-    this.candidates.splice(newIndex, 0, movedCandidate);
+    const movedCandidate = this.expandedCandidates.splice(previousIndex, 1)[0];
+    this.expandedCandidates.splice(newIndex, 0, movedCandidate);
 
-    // First, update the index of all base candidates (isOriginal: true)
-    let updatedIndex = 1;
-    this.candidates.forEach((candidate) => {
-      if (candidate.isOriginal) {
-        candidate.index = updatedIndex;
-        updatedIndex++;
-      }
-    });
-
-    // Now handle clones (isOriginal: false) for each base candidate
-    this.candidates.forEach((candidate) => {
-      if (!candidate.isOriginal) {
-        const baseCandidate = this.candidates.find((c) => c.uuid === candidate.originalUuid);
-
-        if (baseCandidate) {
-          // Update the clone's index to match the current index of its base candidate
-          candidate.index = baseCandidate.index;
-        }
-      }
-    });
-
-    this.reorderCandidates(this.candidates);
+    this.updateCandidatePositions();
     this.updateHasChanges();
   }
 
   private updateCandidatePositions(): void {
-    const processedCandidates = new Set();
+    const processedCandidates = new Set<CandidateModel>();
 
-    for (let i = 1; i <= this.candidates.length; i++) {
-      const currentCandidate = this.candidates[i - 1];
-      const alreadyProcessed = processedCandidates.has(currentCandidate.originalUuid);
-      const candidates = this.candidates.filter((c) => c.originalUuid === currentCandidate.originalUuid);
+    for (let i = 1; i <= this.expandedCandidates.length; i++) {
+      const currentCandidate = this.expandedCandidates[i - 1].candidate;
+      const alreadyProcessed = processedCandidates.has(currentCandidate);
 
-      for (const candidate of candidates) {
-        if (!alreadyProcessed) {
-          candidate.orderIndex = i;
-        } else {
-          candidate.cloneOrderIndex = i;
-        }
-      }
-
-      processedCandidates.add(currentCandidate.originalUuid);
-    }
-  }
-
-  private sortCandidateAscendingWithIndexAndOrderIndex(a: ListCandidateModel, b: ListCandidateModel): number {
-    // Check if both objects have the 'index' property
-    if ('index' in a && 'index' in b) {
-      if (a.index !== b.index) {
-        return a.index - b.index; // Sort by index
+      if (alreadyProcessed) {
+        currentCandidate.cloned = true;
+        currentCandidate.cloneOrderIndex = i;
       } else {
-        return a.orderIndex - b.orderIndex; // If index is same, sort by orderIndex
+        processedCandidates.add(currentCandidate);
+        currentCandidate.cloned = false;
+        currentCandidate.index = i;
+        currentCandidate.orderIndex = i;
       }
-    } else if ('index' in a) {
-      return -1; // 'a' has index, move it before 'b'
-    } else if ('index' in b) {
-      return 1; // 'b' has index, move it before 'a'
-    } else {
-      return a.orderIndex - b.orderIndex; // Neither has index, sort by orderIndex directly
     }
+
+    this.refreshExpandedCandidates();
   }
 
   private refreshExpandedCandidates(): void {
-    const candidateList: ListCandidateModel[] = [];
+    this.expandedCandidates = [];
 
     for (const candidate of this.candidates) {
-      const foundCandidate = candidateList.find((c) => c.id === candidate.id && c.uuid === candidate.uuid);
-      if (foundCandidate) {
-        if (candidate.cloned && foundCandidate.id == candidate.id) {
-          candidateList.push({ ...candidate, orderIndex: candidate.cloneOrderIndex });
-        }
-      } else {
-        candidateList.push(candidate);
+      this.expandedCandidates.push({ candidate, index: candidate.orderIndex - 1, isCloned: false });
+      if (candidate.cloned) {
+        this.expandedCandidates.push({ candidate, index: candidate.cloneOrderIndex - 1, isCloned: true });
       }
     }
-    this.candidates = candidateList;
-    this.dataSource.data = this.candidates;
+
+    this.expandedCandidates.sort((a, b) => a.index - b.index);
+    this.dataSource.data = this.expandedCandidates;
   }
 
   private processCandidates(candidates: CandidateModel[]): void {
-    this.candidates = candidates.map((c) => {
-      const originalUuid = uuid();
-      return {
-        ...c,
-        uuid: originalUuid,
-        originalUuid: originalUuid,
-        isOriginal: true,
-      };
-    });
+    this.initialCandidates = candidates;
     // Create copies of cloned/preCumulated candidates
-    this.candidates
-      .filter((c) => c.cloned)
-      .sort((c1, c2) => c1.cloneOrderIndex - c2.cloneOrderIndex)
-      .forEach((c) => {
-        const clone: ListCandidateModel = {
-          ...c,
-          uuid: uuid(),
-          cloneUuid: c.uuid,
-          isOriginal: false,
-        };
-        this.candidates.splice(clone.cloneOrderIndex - 1, 0, clone);
-        c.cloneUuid = clone.uuid;
-      });
-    this.initialCandidacies = this.candidates.map((c) => ({
+    this.candidates = candidates.map((c) => ({
       ...c,
       markings: [...c.markings],
     }));
-    this.dataSource.data = this.candidates;
   }
 
   private updateHasChanges(): void {
@@ -446,7 +336,7 @@ export class CandidacyOverviewComponent implements OnInit, OnDestroy, AfterViewI
   }
 
   private checkChanges(): boolean {
-    return this.arraysAreDifferent(this.initialCandidacies, this.candidates);
+    return this.arraysAreDifferent(this.initialCandidates, this.candidates);
   }
 
   private arraysAreDifferent(arr1: any[], arr2: any[]): boolean {