+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# ✨ Changelog (`v4.6.2`)
+# ✨ Changelog (`v4.8.1`)
 
 All notable changes to this project will be documented in this file.
 
@@ -8,11 +8,47 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Version Info
 
 ```text
-This version -------- v4.6.2
-Previous version ---- v4.5.10
+This version -------- v4.8.1
+Previous version ---- v4.6.2
 Initial version ----- v4.5.7
-Total commits ------- 5
+Total commits ------- 6
 ```
+
+## [v4.8.1] - 2026-03-04
+
+### 🔄 Changed
+
+- do not show next candidacy button in majority elections with one seat
+
+## [v4.8.0] - 2026-03-04
+
+### 🔄 Changed
+
+- require at least one candidate per list
+
+## [v4.7.1] - 2026-03-04
+
+### 🔄 Changed
+
+- disable drag drop on locked lists
+
+## [v4.7.0] - 2026-03-04
+
+### 🔄 Changed
+
+- remove candidate number in majority elections
+
+## [v4.6.4] - 2026-02-26
+
+### 🔄 Changed
+
+- refactor candidate overview to fix multiple bugs
+
+## [v4.6.3] - 2026-02-25
+
+### 🔄 Changed
+
+- improve archive
 
 ## [v4.6.2] - 2026-02-06